Pull request review fixes
